pyproject.toml
[project]
name = "ctxssg"
version = "0.1.0"
description = "contextual docs"
readme = "README.md"
authors = [{name = "Kyle Gruel"}]
license = "Apache-2.0"
requires-python = ">=3.10"
classifiers = []
dependencies = [
    "click>=8.0",
    "pypandoc>=1.11",
    "pyyaml>=6.0",
    "jinja2>=3.0",
    "watchdog>=3.0",
    "python-frontmatter>=1.0",
    "beautifulsoup4>=4.12",
    "tomli>=2.0.0; python_version<'3.11'",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "black>=24.8.0",
    "mypy>=1.14.1",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.13",
]

[project.urls]
Homepage = "https://github.com/kgruel/ctxssg"
Changelog = "https://github.com/kgruel/ctxssg/releases"
Issues = "https://github.com/kgruel/ctxssg/issues"
CI = "https://github.com/kgruel/ctxssg/actions"

[project.scripts]
ctxssg = "ctxssg.cli:cli"

[project.optional-dependencies]
test = ["pytest"]
